% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-download.R
\name{asset_download}
\alias{asset_download}
\title{Download gblidar assets}
\usage{
asset_download(x, target_dir, resume = FALSE, progress = TRUE, timeout = Inf)
}
\arguments{
\item{x}{A \code{gbl_catalog} object.}

\item{target_dir}{The parent directory declaring where files should be
downloaded}

\item{resume}{if the file already exists, resume the download. Note that
this may change server responses, see details.}

\item{progress}{logical. If TRUE a download progress is printed.}

\item{timeout}{numeric. Default is Inf i.e. download will continue until
completion.}
}
\value{
The function returns a data frame with one row for each downloaded
file and the following columns:
\itemize{
\item success if the HTTP request was successfully performed, regardless of the
response status code. This is FALSE in case of a network error, or in case
you tried to resume from a server that did not support this. A value of NA
means the download was interrupted while in progress.
\item status_code the HTTP status code from the request. A successful download is
usually 200 for full requests or 206 for resumed requests. Anything else
could indicate that the downloaded file contains an error page instead of
the requested content.
\item resumefrom the file size before the request, in case a download was resumed.
\item url final url (after redirects) of the request.
\item destfile downloaded file on disk.
\item error if success == FALSE this column contains an error message.
\item type the Content-Type response header value.
\item modified the Last-Modified response header value.
}

*time total elapsed download time for this file in seconds.
\itemize{
\item headers vector with http response headers for the request.
\item save_location the detination that the file is saved in.
}
}
\description{
Download all assets present in a \code{gbl_catalog} object. This is basically just
a wrapper around \code{curl::multi_download}.
}
