% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{bind_catalogs}
\alias{bind_catalogs}
\title{row bind multiple gbl_catalog objects}
\usage{
bind_catalogs(..., .id)
}
\arguments{
\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use names
if available, otherwise it will use positions.}

\item{...gbl_catalog}{objects to combine}
}
\value{
A filtered gbl_catalog object
}
\description{
wrapper for dplyr::bind_rows for \code{gbl_catalog} objects
}
