% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-wcs.R
\name{eng_composite}
\alias{eng_composite}
\title{Get English Composite LiDAR-derived raster data fast!}
\usage{
eng_composite(
  aoi,
  product = c("fz_dsm", "dsm", "dtm", "vom"),
  product_type = c("elevation", "hillshade"),
  prod_res = 1,
  mask = FALSE,
  destination = NULL,
  warp_res = prod_res,
  resample = c("bilinear", "near", "cubic", "cubicspline", "lanczos", "average", "rms",
    "mode", "max", "min", "med", "q1", "q3"),
  compression = "DEFLATE",
  nodata = NULL,
  options = NULL,
  raster_class = getOption("gblidar.out_raster_type"),
  progress = getOption("gblidar.progress")
)
}
\arguments{
\item{aoi}{A spatial area to request: Can be any of the following classes:
sf, sfc or character. When a character is provided it must be a gdal readable
spatial vector file format.}

\item{product}{character. The product to retrieve. Can be any of the
following: "fz_dsm", "dsm", "dtm", "vom".}

\item{product_type}{character. The type of product to retrieve. Can be any of
the following: "elevation", "hillshade".}

\item{prod_res}{numeric. The resolution of the product. Can be 1 or 2
(in meters).}

\item{mask}{logical. Whether to mask the raster data to the area of interest.}

\item{destination}{character. The destination file path for the raster data.}

\item{warp_res}{numeric. The resolution (in meters) of output raster data.
Can differ from the product resolution (although this is the default).}

\item{resample}{character. The resampling method to use. Can be any of the
following: "near", "bilinear", "cubic", "cubicspline", "lanczos", "average",
"rms", "mode", "max", "min", "med", "q1", "q3".}

\item{compression}{character. The compression method to use. Can be for
example: "DEFLATE" or "LZW".}

\item{nodata}{numeric. The nodata value to use.}

\item{options}{character. A character vector with gdal options.}

\item{raster_class}{character. The class in which to return the raster.
Can be any of "character" (in which case the file path is returned -
the default), "SpatRaster", or "stars".}

\item{progress}{logical. Should a progress bar be shown.}
}
\value{
A \code{character} file path, \code{SpatRaster} or \code{stars} object, depending
on the \code{raster_class} argument.
}
\description{
This function retrieves composite raster data from the
Environment Agency's Web Coverage Service (WCS) for the following products:
"fz_dsm", "dsm", "dtm", "vom".
}
