% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-api.R
\name{eng_post_feature}
\alias{eng_post_feature}
\title{format the area of interest for the Environment Agency API}
\usage{
eng_post_feature(aoi)
}
\arguments{
\item{aoi}{A spatial area to request: Can be any of the following classes:}
}
\description{
format the area of interest for the Environment Agency API
}
