% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{filter_catalog}
\alias{filter_catalog}
\title{filter a gbl_catalog object}
\usage{
filter_catalog(x, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{x}{A gbl_catalog object}
}
\value{
A filtered gbl_catalog object
}
\description{
wrapper for dplyr::filter for \code{gbl_catalog} objects
}
\examples{
scafell_box <- sf::st_point(c(321633, 507181)) |>
  sf::st_buffer(100) |>
  sf::st_sfc() |>
  sf::st_set_crs(27700)

scafell_catalog <- eng_search(scafell_box)

DTM_catalog <- scafell_catalog |>
  filter_catalog(
    product == "LIDAR Composite DTM",
    resolution == 2,
    year == 2022
  )
}
