% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{gbl_paths}
\alias{gbl_paths}
\title{Return the data file paths for a gbl_catalog object}
\usage{
gbl_paths(x)
}
\arguments{
\item{x}{A gbl_catalog object}
}
\value{
a character vector of file paths.
}
\description{
Return the data file paths for a gbl_catalog object
}
