% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{gbl_urls}
\alias{gbl_urls}
\title{Return the data urls for a gbl_catalog object}
\usage{
gbl_urls(x)
}
\arguments{
\item{x}{A gbl_catalog object}
}
\value{
a character vector of url paths.
}
\description{
Return the data urls for a gbl_catalog object
}
