% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{plot.gbl_catalog}
\alias{plot.gbl_catalog}
\title{Plot a gbl_catalog object}
\usage{
\method{plot}{gbl_catalog}(x)
}
\arguments{
\item{x}{a gbl_catalog object}
}
\description{
Plot a gbl_catalog object
}
