% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-wcs.R
\name{wcs_layers}
\alias{wcs_layers}
\title{Web Coverage Service (WCS) layer retrieval}
\usage{
wcs_layers(wcs_url)
}
\arguments{
\item{wcs_url}{character. The WCS endpoint URL from \code{wcs_url()}.}
}
\value{
list. A list of WCS layers.
}
\description{
Web Coverage Service (WCS) layer retrieval
}
