% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng-wcs.R
\name{wcs_url}
\alias{wcs_url}
\title{Web Coverage Service (WCS) endpoint retrieval}
\usage{
wcs_url(product = c("fz_dsm", "dsm", "dtm", "vom"), resolution = 1)
}
\arguments{
\item{product}{character. The product to retrieve. Can be any of the following:
"fz_dsm", "dsm", "dtm", "vom".}

\item{resolution}{numeric. The resolution of the product. Can be 1 or 2.
(in meters)}
}
\value{
object of class \code{eng.wcs.url} and \code{character}. comprising the WCS
endpoint URL.
}
\description{
Web Coverage Service (WCS) endpoint retrieval
}
