% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_catalog.R
\name{filter_catalog}
\alias{filter_catalog}
\title{filter a gbl_catalog object}
\usage{
filter_catalog(x, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{x}{A gbl_catalog object}
}
\value{
A filtered gbl_catalog object
}
\description{
wrapper for dplyr::filter for \code{gbl_catalog} objects
}
