% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset-merge.R
\name{merge_assets}
\alias{merge_assets}
\title{merge raster assets}
\usage{
merge_assets(
  x,
  mask = FALSE,
  destination = NULL,
  resample = "near",
  compression = "DEFLATE",
  nodata = NULL,
  options = NULL,
  raster_class = getOption("gblidar.out_raster_type"),
  progress = getOption("gblidar.progress")
)
}
\arguments{
\item{x}{A \code{gbl_catalog} object}

\item{destination}{out put destination. if NUll then a tempfile is created}

\item{resample}{resampling method, default near}

\item{compression}{the tif compression method to use - e.g. "DEFLATE" or "LZW"}

\item{options}{character vector with gdal options.}

\item{raster_class}{The class in which to return the raster. can be any of
"character" (in which case the file path is returned - the default),
"SpatRaster", or "stars".}

\item{progress}{logical - should a progress bar be shown.}
}
\value{
A raster file path
}
\description{
uses gdal warp to merge raster assets into a single file.
}
